test_log.py: Ensure that py.test -s does not fail

+++ src/test/test_log.py
@@ -25,7 +25,7 @@ def logger():
 
 WARN_FILE_NAME = "test_log_warn.log"
 
-def test_log_redirection(logger, capsys):
+def test_log_redirection(request, logger, capsys):
     # Special handling: Redirect file descriptor 3 output to a file for verification
     capture_warn = py.io.FDCapture(3, open(WARN_FILE_NAME, "wb+"))
     print
@@ -35,11 +35,13 @@ def test_log_redirection(logger, capsys):
     out, err = capsys.readouterr()
     assert out == "\nINFO: logger_info\n"
     assert err == "ERROR: logger_error\n"
-    warn = open(WARN_FILE_NAME).read()
-    assert warn == "WARNING: logger_warn\n"
-    os.unlink(WARN_FILE_NAME)
+    # py.test -s flag undoes the file descriptor redirection from the logger
+    if not request.config.getoption("-s"):
+        warn = open(WARN_FILE_NAME).read()
+        assert warn == "WARNING: logger_warn\n"
+        os.unlink(WARN_FILE_NAME)
 
-def test_log_colors(logger, capsys, monkeypatch):
+def test_log_colors(request, logger, capsys, monkeypatch):
     monkeypatch.setattr(MscBoost.log, "USE_COLORS", True)
     monkeypatch.setattr(MscBoost.log, "FORCE_COLORS", True)
     ESC = chr(27)
@@ -49,9 +51,10 @@ def test_log_colors(logger, capsys, monkeypatch):
     logger.error("logger_error")
     out, err = capsys.readouterr()
     assert err == red+"ERROR: logger_error"+regular+"\n"
-
-    capture_warn = py.io.FDCapture(3, open(WARN_FILE_NAME, "wb+"))
-    logger.warn("logger_warn")
-    warn = open(WARN_FILE_NAME).read()
-    assert warn == yellow+"WARNING: logger_warn"+regular+"\n"
-    os.unlink(WARN_FILE_NAME)
+    # py.test -s flag undoes the file descriptor redirection from the logger
+    if not request.config.getoption("-s"):
+        capture_warn = py.io.FDCapture(3, open(WARN_FILE_NAME, "wb+"))
+        logger.warn("logger_warn")
+        warn = open(WARN_FILE_NAME).read()
+        assert warn == yellow+"WARNING: logger_warn"+regular+"\n"
+        os.unlink(WARN_FILE_NAME)
